+++ CHANGELOG.md
@@ -10,5 +10,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Get all webhooks
+- Get transactions by account
 
 [unreleased]: https://github.com/brainnco/strong_params/compare/main